+++ packages/electron/src/integration-test/integration.test.ts
@@ -37,7 +37,7 @@ const userDataPath = getElectronUserDataPath();
 let electronApp: ElectronApplication;
 let electronPage: Page;
 const testHelper = new TestHelper(signInOptions);
-const tokenStore = new RefreshTokenStore(getTokenStoreFileName(),getTokenStoreKey(), userDataPath);
+const tokenStore = new RefreshTokenStore(getTokenStoreFileName(), getTokenStoreKey(), userDataPath);
 
 function getTokenStoreKey(issuerUrl?: string): string {
   const authority = new URL(issuerUrl ?? "https://ims.bentley.com");
@@ -80,7 +80,6 @@ test.afterEach(async () => {
 });
 
 test("buttons exist", async () => {
-  await electronPage.waitForLoadState("domcontentloaded");
   const signInButton = electronPage.getByTestId("signIn");
   const signOutButton = electronPage.getByTestId("signOut");
   const getStatusButton = electronPage.getByTestId("getStatus");
@@ -93,8 +92,8 @@ test("sign in successful", async ({ browser }) => {
   const page = await browser.newPage();
   await testHelper.checkStatus(electronPage, false);
   await testHelper.clickSignIn(electronPage);
-  await testHelper.signIn(page, await getUrl(electronApp));
-  await page.waitForLoadState("networkidle");
+  const url = await getUrl(electronApp);
+  await testHelper.signIn(page, url);
   await testHelper.checkStatus(electronPage, true);
   await page.close();
 });
@@ -103,10 +102,21 @@ test("sign out successful", async ({ browser }) => {
   const page = await browser.newPage();
   await testHelper.clickSignIn(electronPage);
   await testHelper.signIn(page, await getUrl(electronApp));
-  await page.waitForLoadState("networkidle");
   await testHelper.checkStatus(electronPage, true);
   await testHelper.clickSignOut(electronPage);
-  await page.waitForLoadState("networkidle");
   await testHelper.checkStatus(electronPage, false);
   await page.close();
+});
+
+test("when scopes change, sign in is required", async ({ browser }) => {
+  const page = await browser.newPage();
+  await testHelper.clickSignIn(electronPage);
+  await testHelper.signIn(page, await getUrl(electronApp));
+  await testHelper.checkStatus(electronPage, true);
+
+  // Admittedly this is cheating: no user would interact
+  // with the tokenStore directly, but this is a tough
+  // case to test otherwise.
+  await tokenStore.load("itwin-platform realitydata:read");
+  await testHelper.checkStatus(electronPage, false);
 });