+++ packages/browser/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @itwin/browser-authorization
 
-This log was last generated on Tue, 18 Jun 2024 19:11:49 GMT and should not be manually modified.
+This log was last generated on Wed, 20 Nov 2024 18:34:06 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 1.1.4
+
+Wed, 20 Nov 2024 18:34:06 GMT
+
+### Patches
+
+- Update to @playwright/test 1.48.2 (33036725+wgoehrig@users.noreply.github.com)
+
+## 1.1.3
+
+Tue, 29 Oct 2024 21:49:00 GMT
+
+### Patches
+
+- Fix method name in README.md (GytisCepk@users.noreply.github.com)
 
 ## 1.1.2